sparate material app from main and use screen util package to handle responsive ui
