+++ lib/main.dart
@@ -1,32 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:islamic_app/core/theming/app_theme.dart';
-import 'package:islamic_app/features/splash/splash_screen.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:islamic_app/core/routing/app_routing.dart';
+import 'package:islamic_app/yaqeen_app.dart';
 
-void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: 'Yaqeen',
-      theme: AppTheme.lightMode,
-      //darkTheme: AppTheme.darkMode,
-      //themeMode: ThemeMode.system,
-      localizationsDelegates: const [
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate,
-        GlobalCupertinoLocalizations.delegate,
-      ],
-      supportedLocales: const [
-        Locale('ar'),
-      ],
-      home: const SplashScreen(),
-    );
-  }
+void main() async {
+   WidgetsFlutterBinding.ensureInitialized();
+  await ScreenUtil.ensureScreenSize();
+  runApp(YaqeenApp(
+    appRouter: AppRouting(),
+  ));
 }
